o Fixed PreparedStatementTest.testPrepareFailWarning() to be skipped when run against SQL Server 2008 or above or Sybase ASE.
git-svn-id: http://svn.code.sf.net/p/jtds/code/branches/jTDS%201.3%20(stable)@1151 97a8069c-bbaa-4fa2-9e49-561d8e0b19ef

+++ src/test/net/sourceforge/jtds/jdbc/PreparedStatementTest.java
@@ -817,38 +817,46 @@ public class PreparedStatementTest extends TestBase {
     }
 
     /**
-     * Test for bug [1180777] collation-related execption on update.
-     * <p/>
-     * If a statement prepare fails the statement should still be executed
-     * (unprepared) and a warning should be added to the connection (the
-     * prepare failed, this is a connection event even if it happened on
-     * statement execute).
+     * <p> Test for bug #378, collation-related exception on update. </p>
+     *
+     * <p> If a statement prepare fails the statement should still be executed
+     * (unprepared) and a warning should be added to the connection (the prepare
+     * failed, this is a connection event even if it happened on statement
+     * execute). </p>
      */
-    public void testPrepareFailWarning() throws SQLException {
-        try {
-            PreparedStatement pstmt = con.prepareStatement(
-                    "CREATE VIEW prepFailWarning AS SELECT 1 AS value");
+   public void testPrepareFailWarning() throws SQLException
+   {
+      // preparation succeeds in SQL server 2008 and above (what about Sybase?)
+      if( con.getMetaData().getURL().toLowerCase().contains( "microsoft" ) && con.getMetaData().getDatabaseMajorVersion() < 10 )
+      {
+         try
+         {
+            // FIXME: we need another test case working for newer SQL Server versions and Sybase ASE
+            PreparedStatement pstmt = con.prepareStatement( "CREATE VIEW prepFailWarning AS SELECT 1 AS value" );
             pstmt.execute();
             // Check that a warning was generated on the connection.
             // Although not totally correct (the warning should be generated on
             // the statement) the warning is generated while preparing the
             // statement, so it belongs to the connection.
-            assertNotNull(con.getWarnings());
+            assertNotNull( con.getWarnings() );
             pstmt.close();
 
             Statement stmt = con.createStatement();
-            ResultSet rs = stmt.executeQuery("SELECT * FROM prepFailWarning");
-            assertTrue(rs.next());
-            assertEquals(1, rs.getInt(1));
-            assertFalse(rs.next());
+            ResultSet rs = stmt.executeQuery( "SELECT * FROM prepFailWarning" );
+            assertTrue( rs.next() );
+            assertEquals( 1, rs.getInt( 1 ) );
+            assertFalse( rs.next() );
             rs.close();
             stmt.close();
-        } finally {
+         }
+         finally
+         {
             Statement stmt = con.createStatement();
-            stmt.execute("DROP VIEW prepFailWarning");
+            stmt.execute( "DROP VIEW prepFailWarning" );
             stmt.close();
-        }
-    }
+         }
+      }
+   }
 
     /**
      * Test that preparedstatement logic copes with commit modes and
@@ -1017,7 +1025,7 @@ public class PreparedStatementTest extends TestBase {
         assertEquals(29000000, rs.getBigDecimal(1).intValue());
         stmt.close();
     }
-    
+
     /**
      * Test for bug [1623668] Lost apostrophes in statement parameter values(prepareSQL=0)
      */
@@ -1030,17 +1038,17 @@ public class PreparedStatementTest extends TestBase {
             Statement stmt = con.createStatement();
             stmt.execute("CREATE TABLE #prepareSQL0 (position int, data varchar(32))");
             stmt.close();
-            
+
         	PreparedStatement ps = con.prepareStatement("INSERT INTO #prepareSQL0 (position, data) VALUES (?, ?)");
-        	
+
         	String data1 = "foo'foo";
         	String data2 = "foo''foo";
         	String data3 = "foo'''foo";
-        	
+
         	ps.setInt(1, 1);
         	ps.setString(2, data1);
         	ps.executeUpdate();
-        	
+
         	ps.setInt(1, 2);
         	ps.setString(2, data2);
         	ps.executeUpdate();
@@ -1048,20 +1056,20 @@ public class PreparedStatementTest extends TestBase {
         	ps.setInt(1, 3);
         	ps.setString(2, data3);
         	ps.executeUpdate();
-        	
+
         	ps.close();
         	ps = con.prepareStatement("SELECT data FROM #prepareSQL0 ORDER BY position");
         	ResultSet rs = ps.executeQuery();
-        	
+
         	rs.next();
         	assertEquals(data1, rs.getString(1));
-        	
+
         	rs.next();
         	assertEquals(data2, rs.getString(1));
 
         	rs.next();
         	assertEquals(data3, rs.getString(1));
-        	
+
         	rs.close();
         } finally {
             con.close();
